requirements.txt
flask
anthropic
korean-lunar-calendar
python-dotenv
